UPDATED_PROJECT_FOLDER/requirements.txt
streamlit
pandas
numpy
environ
pyodbc
tabulate
psycopg2-binary
langchain_community
openai
langchain
scikit-learn
matplotlib
seaborn
plotly
opencv-python
pyzbar
python-barcode